schemathesis/requirements.txt
fastapi
mysql-connector-python==8.1.0
pydantic==2.4.2
python-dotenv==1.0.0
starlette
uvicorn
schemathesis
pytest
aiohttp
pytest-asyncio

